+++ common/src/test/java/com/microsoft/identity/common/internal/ui/webview/certbasedauth/TestDialog.java
@@ -1,0 +1,12 @@
+package com.microsoft.identity.common.internal.ui.webview.certbasedauth;
+
+enum TestDialog {
+    cert_picker,
+    pin,
+    error,
+    user_choice,
+    prompt,
+    nfc_loading,
+    nfc_prompt,
+    nfc_reminder
+}
